requirements.txt
dash
protobuf
grpcio
pandas <= 2.0
datetime
python-dateutil
numpy
plotly
requests
flask
gunicorn
nbformat>=4.2.0
openai
python-dotenv
pypandoc